[feat] 디테일뷰 네트워크 연결 상태에 따른 ui 업데이트 구현

+++ app/src/main/java/com/hyeeyoung/wishboard/presentation/wishitem/viewmodels/WishItemViewModel.kt
@@ -8,6 +8,8 @@ import com.hyeeyoung.wishboard.data.model.folder.FolderItem
 import com.hyeeyoung.wishboard.data.model.wish.WishItem
 import com.hyeeyoung.wishboard.domain.model.WishItemDetail
 import com.hyeeyoung.wishboard.domain.repositories.WishRepository
+import com.hyeeyoung.wishboard.presentation.base.viewmodel.NetworkViewModel
+import com.hyeeyoung.wishboard.util.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -18,7 +20,9 @@ import javax.inject.Inject
 @HiltViewModel
 class WishItemViewModel @Inject constructor(
     private val wishRepository: WishRepository,
-) : ViewModel() {
+) : NetworkViewModel() {
+    private val _wishDetailFetchState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
+    val wishDetailFetchState get() = _wishDetailFetchState.asStateFlow()
     private val _wishItemThumbnail = MutableLiveData<WishItem>()
     val wishItemThumbnail: LiveData<WishItem> get() = _wishItemThumbnail
     private val _itemDetail = MutableLiveData<WishItemDetail>()
@@ -32,12 +36,14 @@ class WishItemViewModel @Inject constructor(
             _itemDetail.value =
                 wishRepository.fetchWishItemDetail(itemId)?.map { it.toWishItemDetail() }
                     ?.get(0)
+            _wishDetailFetchState.value = if (itemDetail.value == null) UiState.Error(null) else UiState.Success(true)
             _itemImage.value = itemDetail.value?.image
             generateWishItemThumbnail(itemDetail.value ?: return@launch)
         }
     }
 
     fun deleteWishItem() {
+        if (!isConnected.value) return
         val itemId = itemDetail.value?.id ?: return
         viewModelScope.launch {
             isCompleteDeletion.value = wishRepository.deleteWishItem(itemId)
@@ -45,6 +51,7 @@ class WishItemViewModel @Inject constructor(
     }
 
     fun updateWishItemFolder(folder: FolderItem) {
+        if (!isConnected.value) return
         val item = itemDetail.value?.apply {
             folderId = folder.id
             folderName = folder.name
@@ -82,6 +89,10 @@ class WishItemViewModel @Inject constructor(
             _wishItemThumbnail.value =
                 WishItem(imageUrl = image, name = name, price = price.toIntOrNull(), id = id)
         }
+    }
+
+    fun requestFetchWishDetail() {
+        _wishDetailFetchState.value = UiState.Loading
     }
 
     fun getIsCompleteDeletion(): LiveData<Boolean> = isCompleteDeletion
